initial commit
When using this code in MATLAB, the first step is to upload a clear picture of the microplastic-covered filter paper to MATLAB. This algorithm only works if the only contamination on the filter paper is microplastic; other contamination will also be turned into black pixels and be counted as microplastic. The  threshold level also needs to be adjusted via trial and error. Each image has different amounts of light, so the threshold should be adjusted until the algorithm only turns the microplastic black, not any background. The collected value was the percent of impervious surface calculated.